requirements.txt
coloredlogs==15.0.1
PyYAML==6.0.1
python-telegram-bot[job-queue]==20.5
aio-mc-rcon==3.2.1
